requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6 # Apache-2.0

iso8601>=0.1.9 # MIT
Jinja2>=2.8 # BSD License (3 clause)
keystoneauth1>=2.1.0 # Apache-2.0
os-client-config>=1.13.1 # Apache-2.0
oslo.concurrency>=3.5.0 # Apache-2.0
oslo.config>=3.9.0 # Apache-2.0
oslo.i18n>=2.1.0 # Apache-2.0
oslo.log>=1.14.0 # Apache-2.0
oslo.serialization>=1.10.0 # Apache-2.0
oslo.utils>=3.5.0 # Apache-2.0
psutil<2.0.0,>=1.1.1 # BSD
pygal
pykwalify
python-glanceclient>=2.0.0 # Apache-2.0
python-neutronclient>=4.2.0 # Apache-2.0
python-novaclient!=2.33.0,>=2.29.0 # Apache-2.0
python-heatclient>=0.6.0 # Apache-2.0
python-subunit>=0.0.18 # Apache-2.0/BSD
PyYAML>=3.1.0 # MIT
pyzmq>=15.2.0
six>=1.9.0 # MIT
hdrhistogram>=0.5.2
